fix: ensure remove events on reload page #134
