--- lib/core/event-manager.ts
@@ -1,6 +1,0 @@
-export default abstract class EventManager {
-    abstract subscribe(eventName: string, fn: (...args: any[]) => void | Promise<void>): void;
-    abstract exists(eventName: string): boolean;
-    abstract removerEvent(eventName: string): boolean;
-    abstract dispatchEvent(eventName: string, ...args: any[]): void;
-}
