+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "rich-domain",
-	"version": "1.20.3-beta.0",
+	"version": "1.20.3-beta.1",
 	"description": "This package provide utils file and interfaces to assistant build a complex application with domain driving design",
 	"main": "index.js",
 	"types": "index.d.ts",
